Use same names as the ones registered in Angular
The directive ngeoSearch got the name search.js.
The value ngeoCreateGeoJSONBloodhound got the name creategeojsonbloodhound.js.
The module ngeoSearchModule got the name searchmodule.js.

This diminishes brain pressure at the price of even longer names.
